+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Submitting jobs using `Submitter.submit` is now thread-safe.
 - Handling exceptions that occur when reading a configuration file.
 - When reading configuration file fails, an exception is no longer raised, instead an error is reported and a default configuration is used.
+- `CFG` is now a frozen dataclass.
 
 ***